Service layer centralization — Phase 2A: split the CorpusObjsService monolith (#1716)
Split the ~2,900-line corpus_objs_service.py monolith into a segmented
corpuses/services/ package, one cohesive module per responsibility, each
inheriting the Phase 1 BaseService:

- paths.py            CorpusPathService       DocumentPath disambiguation
- corpus_documents.py CorpusDocumentService   doc-in-corpus R/W + membership
- lifecycle.py        DocumentLifecycleService soft-delete / restore / trash
- folders.py          FolderService           folder CRUD + doc-in-folder

Pure, behaviour-preserving relocation: all 40 methods move byte-for-byte
(verified against the original); the only code change is 14 cross-module
call sites rewritten from the monolith's cls.<helper> to an explicit
owning-service reference. corpus_objs_service.py becomes a thin shim whose
CorpusObjsService facade multiply-inherits the four services, so every
existing call site and the test_corpus_objs_service.py regression suite
keep working — the suite's only edits are 9 @patch target relocations
following moved symbols.

Adds test_corpus_services_package.py covering the structural contract
(package layout, BaseService inheritance, facade MRO, standalone service
operation, cross-service delegation) and the Phase 2 implementation plan.

Phase 2 of the service-layer centralization roadmap; Phases 2B
(corpus_service.py for Corpus-row CRUD) and 2C (caller migration + shim
deletion) follow.
