+++ CHANGELOG.md
@@ -26,6 +26,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Service-layer centralization — Phase 2A: split the `CorpusObjsService` monolith into the `corpuses/services/` package** (issue #1716; `opencontractserver/corpuses/services/` — new `__init__.py`, `paths.py`, `corpus_documents.py`, `lifecycle.py`, `folders.py`; `opencontractserver/corpuses/corpus_objs_service.py`; `opencontractserver/tests/test_corpus_objs_service.py`; new `opencontractserver/tests/test_corpus_services_package.py`; plan in `docs/refactor_plans/2026-05-21-service-layer-phase2-corpus-services-plan.md`). Phase 2 of the service-layer centralization roadmap (`docs/refactor_plans/2026-05-19-service-layer-centralization-design.md` §6), building on the Phase 1 `BaseService` foundation (#1715). The ~2,920-line `corpus_objs_service.py` monolith — a single `CorpusObjsService` class holding six distinct responsibilities — is split into a segmented `corpuses/services/` package, one cohesive module per responsibility, each service inheriting `BaseService`:
+  - `paths.py` — `CorpusPathService` (6 methods): low-level `DocumentPath` disambiguation internals (`_compute_moved_path`, `_disambiguate_path`, `_create_successor_path_with_retry`, …). All private; no permission checks.
+  - `corpus_documents.py` — `CorpusDocumentService` (13 methods): document-in-corpus reads/writes and membership — including the two deliberate `get_corpus_documents` (corpus-as-gate) / `get_corpus_documents_visible_to_user` (MIN-permission) semantics from #1682.
+  - `lifecycle.py` — `DocumentLifecycleService` (5 methods): soft-delete / restore / trash.
+  - `folders.py` — `FolderService` (16 methods): folder CRUD, the folder tree, and document-in-folder placement.
+  - **Pure relocation — no behaviour change.** All 40 methods are moved byte-for-byte; permission checks, transaction boundaries, query shapes, and logging are untouched. The only code change is **14 cross-module call sites** (12 in `folders.py`, 2 in `lifecycle.py`) rewritten from the monolith's `cls.<helper>` to an explicit owning-service reference (`CorpusPathService._disambiguate_path(...)`, `CorpusDocumentService._check_document_in_corpus(...)`); within-module calls keep `cls.`. An automated faithfulness check confirms every relocated method body is byte-identical to the monolith modulo those 14 rewrites.
+  - **Backward compatibility.** `corpus_objs_service.py` is now a ~75-line shim: it re-exports the four services and defines `CorpusObjsService` as a deprecated facade that multiply-inherits all four. The four services are flat `BaseService` subclasses with disjoint method names, so the facade's C3 MRO is unambiguous and every `CorpusObjsService.<method>` call still resolves. All ~37 caller files (`config/graphql/`, `mcp/`, `llms/`, `tasks/`, `discovery/`, …) keep working with no change; caller migration and shim deletion are deferred to Phase 2C.
+  - **Tests.** `test_corpus_objs_service.py` (290+ tests) remains the behavioural regression net — run unchanged through the facade — with the sole edit being 9 `@patch` / `patch.object` target relocations following moved symbols (`post_save` → `services.paths`, `_disambiguate_path` → `CorpusPathService`) plus one import. New `test_corpus_services_package.py` (24 tests) covers the structural contract: package layout, `BaseService` inheritance, the facade's MRO / method aggregation, standalone operation of each segmented service, and cross-service delegation.
+
 - **Permissioning documentation refresh — `user_can` is the canonical authorization API** (issue #1660; `docs/permissioning/consolidated_permissioning_guide.md`, `CLAUDE.md`, `mkdocs.yml`). Documents the multi-phase permission-centralization outcome:
   - **`Manager.user_can` / `obj.user_can` API** (shipped by PR #1663). The canonical single-object authorization check, paired with the `visible_to_user` queryset filter and pinned to agree by `test_authorization_invariants.py`. The consolidated guide's structural-enforcement banner, "single source of truth" section, inheritance references, overview tables, and Agent/LLM section now point at `AnnotationManager.user_can` / `RelationshipManager.user_can` (`opencontractserver/shared/Managers.py`) instead of `user_has_permission_for_obj` at stale `permissioning.py` line numbers. New "The `user_can` Authorization API" section added with a checklist for extending `user_can` to a new visibility-managed model.
   - **Two-tier permission cache** (PR #1665). New guide subsection: Tier-1 per-instance memoization (always on, used by Celery/agent/internal code) vs. Tier-2 request-scoped optimizer (opt-in via `request=info.context`). Convention: pass `request=` in GraphQL resolvers/mutations, omit it everywhere else.
