+++ opencontractserver/corpuses/services/__init__.py
@@ -1,0 +1,39 @@
+"""Corpus-scoped service layer.
+
+Segmented services for corpus-scoped object access and permissioning, split
+out of the former ``corpus_objs_service.py`` monolith (issue #1716,
+service-layer centralization Phase 2 — see
+``docs/refactor_plans/2026-05-21-service-layer-phase2-corpus-services-plan.md``).
+Each service inherits :class:`opencontractserver.shared.services.base.BaseService`.
+
+- :class:`~opencontractserver.corpuses.services.folders.FolderService`
+  — folder CRUD, the folder tree, and document-in-folder placement.
+- :class:`~opencontractserver.corpuses.services.corpus_documents.CorpusDocumentService`
+  — document-in-corpus reads / writes and corpus membership.
+- :class:`~opencontractserver.corpuses.services.lifecycle.DocumentLifecycleService`
+  — soft-delete / restore / trash.
+- :class:`~opencontractserver.corpuses.services.paths.CorpusPathService`
+  — low-level :class:`DocumentPath` disambiguation internals.
+
+Import the specific service you need from this package::
+
+    from opencontractserver.corpuses.services import FolderService
+
+The legacy ``CorpusObjsService`` facade in
+``opencontractserver.corpuses.corpus_objs_service`` is deprecated and will be
+removed once all call sites are migrated.
+"""
+
+from opencontractserver.corpuses.services.corpus_documents import (
+    CorpusDocumentService,
+)
+from opencontractserver.corpuses.services.folders import FolderService
+from opencontractserver.corpuses.services.lifecycle import DocumentLifecycleService
+from opencontractserver.corpuses.services.paths import CorpusPathService
+
+__all__ = [
+    "FolderService",
+    "CorpusDocumentService",
+    "DocumentLifecycleService",
+    "CorpusPathService",
+]
