+++ opencontractserver/tests/test_corpus_objs_service.py
@@ -34,6 +34,7 @@ from opencontractserver.corpuses.models import (
     Corpus,
     CorpusFolder,
 )
+from opencontractserver.corpuses.services import CorpusPathService
 from opencontractserver.documents.document_service import DocumentService
 from opencontractserver.documents.models import Document, DocumentPath
 from opencontractserver.types.enums import PermissionTypes
@@ -2508,7 +2509,7 @@ class TestDocumentPathHistory_DeleteFolderTracking(_DocumentPathHistoryTestBase)
         )
         self.assertEqual(current.path, "/summary.pdf")
 
-    @patch("opencontractserver.corpuses.corpus_objs_service.post_save")
+    @patch("opencontractserver.corpuses.services.paths.post_save")
     def test_delete_folder_dispatches_post_save_for_each_created_path(
         self, mock_signal
     ):
@@ -2649,7 +2650,7 @@ class TestDocumentPathHistory_BulkMoveTracking(_DocumentPathHistoryTestBase):
             DocumentPath.objects.filter(document=doc2, corpus=self.corpus).count(), 1
         )
 
-    @patch("opencontractserver.corpuses.corpus_objs_service.post_save")
+    @patch("opencontractserver.corpuses.services.paths.post_save")
     def test_bulk_move_dispatches_post_save_for_each_created_path(self, mock_signal):
         """bulk_create bypasses Django signals; verify manual dispatch fires once per doc."""
         doc1, _ = self._create_doc_at_root("Doc 1", "doc1.pdf")
@@ -3138,7 +3139,7 @@ class TestErrorPaths_DeleteFolderAtomicRollback(_CorpusObjsServiceFolderTestBase
         )
 
         with patch.object(
-            CorpusObjsService,
+            CorpusPathService,
             "_disambiguate_path",
             side_effect=ValueError("all suffixes exhausted"),
         ):
@@ -3204,7 +3205,7 @@ class TestErrorPaths_DeleteFolderAtomicRollback(_CorpusObjsServiceFolderTestBase
             return original_disambiguate(base_path, corpus, **kwargs)
 
         with patch.object(
-            CorpusObjsService,
+            CorpusPathService,
             "_disambiguate_path",
             staticmethod(fail_on_second),
         ):
@@ -3256,7 +3257,7 @@ class TestErrorPaths_DeleteFolderAtomicRollback(_CorpusObjsServiceFolderTestBase
 
         # First attempt: fails
         with patch.object(
-            CorpusObjsService,
+            CorpusPathService,
             "_disambiguate_path",
             side_effect=ValueError("temporary failure"),
         ):
@@ -3307,7 +3308,7 @@ class TestErrorPaths_DeleteFolderAtomicRollback(_CorpusObjsServiceFolderTestBase
         )
 
         with patch.object(
-            CorpusObjsService,
+            CorpusPathService,
             "_disambiguate_path",
             side_effect=ValueError("blocked"),
         ):
@@ -3408,7 +3409,7 @@ class TestErrorPaths_MoveDocumentIntegrityError(_CorpusObjsServiceFolderTestBase
     def test_disambiguate_exhaustion_returns_error(self):
         """ValueError from _disambiguate_path is surfaced as a user-facing error."""
         with patch.object(
-            CorpusObjsService,
+            CorpusPathService,
             "_disambiguate_path",
             side_effect=ValueError("all suffixes exhausted"),
         ):
@@ -3478,7 +3479,7 @@ class TestErrorPaths_BulkMoveAtomicRollback(_CorpusObjsServiceFolderTestBase):
             return original_disambiguate(base_path, corpus, **kwargs)
 
         with patch.object(
-            CorpusObjsService, "_disambiguate_path", staticmethod(selective_fail)
+            CorpusPathService, "_disambiguate_path", staticmethod(selective_fail)
         ):
             moved_count, error = CorpusObjsService.move_documents_to_folder(
                 user=self.owner,
@@ -3525,7 +3526,7 @@ class TestErrorPaths_BulkMoveAtomicRollback(_CorpusObjsServiceFolderTestBase):
 
         # First attempt: fails
         with patch.object(
-            CorpusObjsService,
+            CorpusPathService,
             "_disambiguate_path",
             side_effect=ValueError("temporary failure"),
         ):
